requirements.txt
PyYAML~=5.4.1
coloring~=4.7.2
blob~=1.0.1
numpy~=1.22.3
unicorn~=0.14.0
fast~=0.68.2
pedantic~=1.9.0
starlet~=0.14.2
tenacity~=8.0.1
requests~=2.27.1
websockets~=9.1
pandas~=1.1.5
setuptools~=54.2.0
scikit-learn~=1.0.2
coloring~=4.7.2
pip~=21.1.2
blob~=1.0.1
tensorflow~=2.8.0
unicorn~=0.14.0
fast~=0.68.2
pedantic~=1.9.0
starlet~=0.14.2
tabulate~=0.8.9